Rename ulmfit to multfit to make room for future directions of the repo.
