+++ multifit/configurations.py
