+++ prepare_imdb.sh
@@ -7,5 +7,5 @@ echo "Saving data in $DATA_DIR"
 wget -c "http://files.fast.ai/data/aclImdb.tgz" -P "${DATA_DIR}"
 
 echo "Imdb is raw text no preparation is done"
-python -m ulmfit.datasets.utils prepare_imdb "${DATA_DIR}/aclImdb.tgz"
+python -m multifit.datasets.utils prepare_imdb "${DATA_DIR}/aclImdb.tgz"
